Refactor service class location
