+++ user-interface/src/main/java/life/qbic/datamanager/views/projects/project/info/ProjectSummaryComponent.java
@@ -61,12 +61,13 @@ import life.qbic.datamanager.views.projects.ProjectInformation;
 import life.qbic.datamanager.views.projects.edit.EditContactDialog;
 import life.qbic.datamanager.views.projects.edit.EditFundingInformationDialog;
 import life.qbic.datamanager.views.projects.edit.EditProjectDesignDialog;
-import life.qbic.datamanager.views.projects.project.info.AsyncProjectService.AccessDeniedException;
-import life.qbic.datamanager.views.projects.project.info.AsyncProjectService.ProjectDesign;
-import life.qbic.datamanager.views.projects.project.info.AsyncProjectService.ProjectUpdateRequest;
-import life.qbic.datamanager.views.projects.project.info.AsyncProjectService.ProjectUpdateResponse;
-import life.qbic.datamanager.views.projects.project.info.AsyncProjectService.RequestFailedException;
-import life.qbic.datamanager.views.projects.project.info.AsyncProjectService.UnknownRequestException;
+import life.qbic.projectmanagement.application.api.AsyncProjectService;
+import life.qbic.projectmanagement.application.api.AsyncProjectService.AccessDeniedException;
+import life.qbic.projectmanagement.application.api.AsyncProjectService.ProjectDesign;
+import life.qbic.projectmanagement.application.api.AsyncProjectService.ProjectUpdateRequest;
+import life.qbic.projectmanagement.application.api.AsyncProjectService.ProjectUpdateResponse;
+import life.qbic.projectmanagement.application.api.AsyncProjectService.RequestFailedException;
+import life.qbic.projectmanagement.application.api.AsyncProjectService.UnknownRequestException;
 import life.qbic.datamanager.views.strategy.dialog.ClosingWithWarningStrategy;
 import life.qbic.datamanager.views.strategy.dialog.ImmediateClosingStrategy;
 import life.qbic.datamanager.views.strategy.scope.ReadScopeStrategy;
